feat(views.py): purge_spammer_account_data no longer deletes sheets -- rather makes unlisted and sets spam_sheet_quarantine to None https://trello.com/c/WW3YdYHv

+++ sefaria/views.py
@@ -941,14 +941,18 @@ def profile_spam_dashboard(request):
         })
 
 
-def purge_spammer_account_data(spammer_id):
+def purge_spammer_account_data(spammer_id, delete_from_nationbuilder=True):
     from django.contrib.auth.models import User
     # Delete from Nationbuilder
     profile = db.profiles.find_one({"id": spammer_id})
-    if "nationbuilder_id" in profile:
+    if delete_from_nationbuilder and "nationbuilder_id" in profile:
         delete_from_nationbuilder_if_spam(spammer_id, profile["nationbuilder_id"])
-    # Delete Sheets
-    db.sheets.delete_many({"owner": spammer_id})
+    sheets = db.sheets.find({"owner": spammer_id})
+    for sheet in sheets:
+        sheet["spam_sheet_quarantine"] = datetime.now()
+        sheet["datePublished"] = None
+        sheet["displayedCollection"] = None
+        db.sheets.save(sheet)
     # Delete Notes
     db.notes.delete_many({"owner": spammer_id})
     # Delete Notifcations
@@ -962,6 +966,7 @@ def purge_spammer_account_data(spammer_id):
     spammer_account = User.objects.get(id=spammer_id)
     spammer_account.is_active = False
     spammer_account.save()
+
 
 @staff_member_required
 def spam_dashboard(request):
